# medical-dictation-backend/docs/api/report-service-openapi.json
{
  "components": {
    "schemas": {
...
      "CreateReportRequest": {
        "additionalProperties": false,
        "properties": {
          "co_author_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Co Author Ids",
            "type": "array"
          },
          "content": {
            "$ref": "#/components/schemas/ReportContent"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "source_session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Session Id"
          }
        },
        "required": [
          "content",
          "patient_id"
        ],
        "title": "CreateReportRequest",
        "type": "object"
      },
...
      "Icd10Code": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 8,
            "minLength": 3,
            "pattern": "^[A-Z][0-9]{2}(\\.[0-9A-Z]{1,4})?$",
            "title": "Code",
            "type": "string"
          },
          "display": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display"
          }
        },
        "required": [
          "code"
        ],
        "title": "Icd10Code",
        "type": "object"
      },
...
      "ReportContent": {
        "additionalProperties": false,
        "description": "The full content_jsonb body for one version.",
        "properties": {
          "encounter_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Date"
          },
          "icd10_codes": {
            "items": {
              "$ref": "#/components/schemas/Icd10Code"
            },
            "title": "Icd10 Codes",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "template_schema_version": {
            "minimum": 1.0,
            "title": "Template Schema Version",
            "type": "integer"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "template_id",
          "template_schema_version"
        ],
        "title": "ReportContent",
        "type": "object"
      },
      "ReportCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "version_id",
          "version_number",
          "status"
        ],
        "title": "ReportCreatedResponse",
        "type": "object"
      },
...
      "ReportSection": {
        "additionalProperties": false,
        "properties": {
          "field_specific_metadata": {
            "additionalProperties": true,
            "title": "Field Specific Metadata",
            "type": "object"
          },
          "icd10": {
            "items": {
              "$ref": "#/components/schemas/Icd10Code"
            },
            "title": "Icd10",
            "type": "array"
          },
          "section_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Section Key",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "transcript_segment_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Transcript Segment Ids",
            "type": "array"
          }
        },
        "required": [
          "section_key"
        ],
        "title": "ReportSection",
        "type": "object"
      },
...
      "UpdateDraftRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/ReportContent"
          },
          "dictation_session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dictation Session Id"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "content"
        ],
        "title": "UpdateDraftRequest",
        "type": "object"
      },
      "UpdateDraftResponse": {
        "additionalProperties": false,
        "properties": {
          "diff_summary": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Diff Summary",
            "type": "object"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "version_id",
          "version_number",
          "status",
          "diff_summary"
        ],
        "title": "UpdateDraftResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Sprint-06 templates + sprint-08 reports / versions / diff / search.",
    "title": "Report Service",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reports": {
      "post": {
        "operationId": "create_report_v1_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "`patient_not_found`, or sprint-13 field-metadata validation: `field_metadata_invalid` / `choice_value_unknown` (section-addressed problems in `problems[]`)."
          }
        },
        "summary": "Create Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/v1/reports/{report_id}/draft": {
      "put": {
        "operationId": "update_draft_v1_reports__report_id__draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Sprint-13 field-metadata validation failed: `field_metadata_invalid` (unknown keys / wrong shape / metadata on a field type that accepts none) or `choice_value_unknown` (a `selected` value not among the section's template options). Section-addressed problems in `problems[]`."
          }
        },
        "summary": "Update Draft",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}